Result: adjust null check performance; refactor
