+++ src/core-taggeds-result/Result/Result.T/Result.T.Fold.cs
@@ -10,29 +10,35 @@ namespace System
         public TResult Fold<TResult>(
             Func<TSuccess, TResult> mapSuccess,
             Func<TFailure, TResult> mapFailure)
-            =>
-            Union.Fold(
-                mapSuccess ?? throw new ArgumentNullException(nameof(mapSuccess)),
-                mapFailure ?? throw new ArgumentNullException(nameof(mapFailure)),
-                Fold_OtherFactory_MustNeverBeInvoked<TResult>);
+        {
+            _ = mapSuccess ?? throw new ArgumentNullException(nameof(mapSuccess));
+            _ = mapFailure ?? throw new ArgumentNullException(nameof(mapFailure));
+
+            return Union.Fold(
+                mapSuccess, mapFailure, Fold_OtherFactory_MustNeverBeInvoked<TResult>);
+        }
 
         public Task<TResult> FoldAsync<TResult>(
             Func<TSuccess, Task<TResult>> mapSuccessAsync,
             Func<TFailure, Task<TResult>> mapFailureAsync)
-            =>
-            Union.FoldAsync(
-                mapSuccessAsync ?? throw new ArgumentNullException(nameof(mapSuccessAsync)),
-                mapFailureAsync ?? throw new ArgumentNullException(nameof(mapFailureAsync)),
-                Fold_OtherFactory_MustNeverBeInvoked<TResult>);
+        {
+            _ = mapSuccessAsync ?? throw new ArgumentNullException(nameof(mapSuccessAsync));
+            _ = mapFailureAsync ?? throw new ArgumentNullException(nameof(mapFailureAsync));
+
+            return Union.FoldAsync(
+                mapSuccessAsync, mapFailureAsync, Fold_OtherFactory_MustNeverBeInvoked<TResult>);
+        }
 
         public ValueTask<TResult> FoldValueAsync<TResult>(
             Func<TSuccess, ValueTask<TResult>> mapSuccessAsync,
             Func<TFailure, ValueTask<TResult>> mapFailureAsync)
-            =>
-            Union.FoldValueAsync(
-                mapSuccessAsync ?? throw new ArgumentNullException(nameof(mapSuccessAsync)),
-                mapFailureAsync ?? throw new ArgumentNullException(nameof(mapFailureAsync)),
-                Fold_OtherFactory_MustNeverBeInvoked<TResult>);
+        {
+            _ = mapSuccessAsync ?? throw new ArgumentNullException(nameof(mapSuccessAsync));
+            _ = mapFailureAsync ?? throw new ArgumentNullException(nameof(mapFailureAsync));
+
+            return Union.FoldValueAsync(
+                mapSuccessAsync, mapFailureAsync, Fold_OtherFactory_MustNeverBeInvoked<TResult>);
+        }
 
         [DoesNotReturn]
         private static TResult Fold_OtherFactory_MustNeverBeInvoked<TResult>()
