+++ src/core-taggeds-result/Result/Result.T/Result.T.MapSuccess.cs
